revert changes

+++ mm/vmpressure.c
@@ -536,7 +536,7 @@ int vmpressure_register_event(struct mem_cgroup *memcg,
 	/* Find required level */
 	token = strsep(&spec, ",");
 	level = str_to_level(token);
-	if ((int)level == -1) {
+	if (level == -1) {
 		ret = -EINVAL;
 		goto out;
 	}
@@ -545,7 +545,7 @@ int vmpressure_register_event(struct mem_cgroup *memcg,
 	token = strsep(&spec, ",");
 	if (token) {
 		mode = str_to_mode(token);
-		if ((int)mode == -1) {
+		if (mode == -1) {
 			ret = -EINVAL;
 			goto out;
 		}
